LyapunovExponents

Documentation for LyapunovExponents.

LyapunovExponents.LyapunovExponents

LyapunovExponents.Network

LyapunovExponents.dimension

LyapunovExponents.eigenmodes

LyapunovExponents.isstable

LyapunovExponents.lyapunovs

LyapunovExponents.msf

LyapunovExponents.solvenet

LyapunovExponents.stablemodes

LyapunovExponents.unstablemodes

LyapunovExponents.LyapunovExponents — Module

A module to calculate lyapunov exponents.

LyapunovExponents.Network — Type

struct Network{T1<:LyapunovExponents.Dynamics, T2<:(AbstractMatrix{T} where T), T3<:(AbstractMatrix{T} where T)}

Network of dynamical system nodes.

Fields

nodes::Vector{T1} where T1<:LyapunovExponents.Dynamics

Dynamical system nodes

E::AbstractMatrix{T} where T

Outer coupling matrix

P::AbstractMatrix{T} where T

Inner coupling matrix

LyapunovExponents.dimension — Method

dimension(ds)

Returns the dimension of ds

LyapunovExponents.eigenmodes — Method

eigenmodes(net)

Returns eigenmodes, i.e. a vector of pairs of eigenvalues of the outer coupling matrix of net and corresponding maximum Lyapunov exponent.

LyapunovExponents.isstable — Method

isstable(net)

Returns true if net is stable. Stability analyis is performed via master stability function.

LyapunovExponents.lyapunovs — Method

lyapunovs(ds; kwargs...)

Computes lyapunov exponents of ds.kwargs may include

nsteps::Int = 3e4 Number of steps to calculate lypaunov exponents

ntrsteps::Int = 0 Number of transient steps for transient steps. Transient steps are taken before calculation of lypaunov exponents

dt::Real = 0.01 Step size of integrator.

LyapunovExponents.msf — Method

msf(ds, λ, P; kwargs...)

Master stability function (msf). Returns maximum lyapunov exponent. kwargs may include

nsteps::Int = 3e4 Number of steps to calculate lypaunov exponents

ntrsteps::Int = 0 Number of transient steps for transient steps. Transient steps are taken before calculation of lypaunov exponents

dt::Real = 0.01 Step size of integrator.

LyapunovExponents.solvenet — Function

solvenet(net, tspan)
solvenet(net, tspan, alg; solkwargs...)

Solves the network differential equation

LyapunovExponents.stablemodes — Method

stablemodes(net)

Returns stable eigen modes of net.

LyapunovExponents.unstablemodes — Method

unstablemodes(net)

Returns unstable eigen modes of net.