LyapunovExponents

Documentation for LyapunovExponents.

LyapunovExponents.LyapunovExponents

LyapunovExponents.Network

LyapunovExponents.addlink!

LyapunovExponents.deletelink!

LyapunovExponents.dimension

LyapunovExponents.eigenmodes

LyapunovExponents.haslink

LyapunovExponents.isadjacency

LyapunovExponents.islaplacian

LyapunovExponents.isstable

LyapunovExponents.lyapunovs

LyapunovExponents.msf

LyapunovExponents.rewirelink!

LyapunovExponents.scale_connection_matrix!

LyapunovExponents.solvenet

LyapunovExponents.stablemodes

LyapunovExponents.unstablemodes

LyapunovExponents.updatelink!

LyapunovExponents.LyapunovExponents — Module

A module to calculate lyapunov exponents.

LyapunovExponents.Network — Type

struct Network{T1<:LyapunovExponents.Dynamics, T2<:(AbstractMatrix{T} where T), T3<:(AbstractMatrix{T} where T)}

Network of dynamical system nodes.

Fields

nodes::Vector{T1} where T1<:LyapunovExponents.Dynamics

Dynamical system nodes

E::AbstractMatrix{T} where T

Outer coupling matrix

P::AbstractMatrix{T} where T

Inner coupling matrix

LyapunovExponents.addlink! — Function

addlink!(net, i, j)
addlink!(net, i, j, w)

Adds a link between the nodes i and j with weight w.

LyapunovExponents.deletelink! — Method

deletelink!(net, i, j)

Deletes the links between the nodes i and j

LyapunovExponents.dimension — Method

dimension(ds)

Returns the dimension of ds

LyapunovExponents.eigenmodes — Method

eigenmodes(net)

Returns eigenmodes, i.e. a vector of pairs of eigenvalues of the outer coupling matrix of net and corresponding maximum Lyapunov exponent.

LyapunovExponents.haslink — Method

haslink(net, i, j)

Returns true if there exists a link between the nodes i and j.

LyapunovExponents.isadjacency — Method

isadjacency(net)

Returns true if net is constructed using adjacency matrix

LyapunovExponents.islaplacian — Method

islaplacian(net)

Returns true if net is constructed using laplacian matrix

LyapunovExponents.isstable — Method

isstable(net)

Returns true if net is stable. Stability analyis is performed via master stability function.

LyapunovExponents.lyapunovs — Method

lyapunovs(ds; kwargs...)

Computes lyapunov exponents of ds.kwargs may include

nsteps::Int = 3e4 Number of steps to calculate lypaunov exponents

ntrsteps::Int = 0 Number of transient steps for transient steps. Transient steps are taken before calculation of lypaunov exponents

dt::Real = 0.01 Step size of integrator.

LyapunovExponents.msf — Method

msf(ds, η, P; kwargs...)

Master stability function (msf). Returns maximum lyapunov exponent of the network defined as

\[\dot{x}_i = f(x)_i + \sum_{j = 1}^n \epsilon_{ij} P x_j \quad i = 1, 2, \ldots, n\]

where $x_i \in \mathbb{R}^d$ is the state vector of the $i^{th}$ node, $f$ is the vector-valued function defining the individual node dynamics. $\epsilon_{ij} \geq 0$ if there is a connection between the nodes $i$ and $j$. $\epsilon_{ij} = 0$, if there is no connection between the nodes $i$ and $j$. $E = [\epsilon_{ij}]$ determines he network topology. $P = diag(p_1, p_2, \ldots, p_d)$ determines the variables by which the nodes are conected to each other. The network can be expressed more compaclty as

\[\dot{X} = F(X) + (E \otimes P) X \]

where $X = [x_1, x_2, \ldots, x_n], F(X) = [f(x_1), f_(x_2), \ldots, f_(x_n)], E = [\epsilon_{ij}], P = diag(p_1, p_2, \ldots, p_n)$

kwargs are

nsteps::Int = 3e4 Number of steps to calculate lypaunov exponents

ntrsteps::Int = 0 Number of transient steps for transient steps. Transient steps are taken before calculation of lypaunov exponents

dt::Real = 0.01 Step size of integrator.

Note

Note that $\epsilon_{ij} \geq 0, j \neq i`,$E$has nonpositive eigenvalues. Let$\eta$denote the eigenvalue of$E``, than we have,

\[ \nu_n \leq \nu_{n - 1} \leq \ldots \leq \nu_2 < \nu_1 = 0\]

LyapunovExponents.rewirelink! — Method

rewirelink!(net, i, j, k, l)

Removes the link from between the nodes i and j and wires the link between the nodes k and l.

LyapunovExponents.scale_connection_matrix! — Method

scale_connection_matrix!(W0)

Normalize the connection matrix of net. The connectin matrix is rescaled such that on-diagonal elements becones 1.

LyapunovExponents.solvenet — Function

solvenet(net, tspan)
solvenet(net, tspan, alg; solkwargs...)

Solves the network differential equation

LyapunovExponents.stablemodes — Method

stablemodes(net)

Returns stable eigen modes of net.

LyapunovExponents.unstablemodes — Method

unstablemodes(net)

Returns unstable eigen modes of net.

LyapunovExponents.updatelink! — Function

updatelink!(net, i, j)
updatelink!(net, i, j, w)

Addes a link between the nodes i and j of the network net. w is the weight of the link to be updated.